Fix regressions related to LightUpdateDispatcher scheduling introduced by https://github.com/Euphillya/Fidorial/pull/85
Also optimized other pieces of code while was at it

+++ fidorial-server/src/main/java/fr/euphyllia/fidorial/server/world/light/LongIntQueue.java
@@ -7,7 +7,7 @@ final class LongIntQueue {
     private int[] xs;
     private int[] ys;
     private int[] zs;
-    private int[] levels;
+    private byte[] levels;
     private int head;
     private int tail;
 
@@ -20,7 +20,7 @@ final class LongIntQueue {
         xs = new int[cap];
         ys = new int[cap];
         zs = new int[cap];
-        levels = new int[cap];
+        levels = new byte[cap];
     }
 
     void reset() {
@@ -35,7 +35,7 @@ final class LongIntQueue {
         xs[tail] = x;
         ys[tail] = y;
         zs[tail] = z;
-        levels[tail] = level;
+        levels[tail] = (byte) level;
         tail++;
     }
